Improve expansion parsing by not discarding unmarshal errors
As described in #570, the current pattern in custom `UnmarshalJSON`
implementations used to perform object expansion is quite poor because
it assumes that *any* unmarshal error on an object indicates that the
object was actually a string. This is not true, and these custom
functions can have the effect of hiding actual problems and making
debugging difficult (see #569).

Here I introduce a new pattern for custom `UnmarshalJSON`
implementations. Instead of trying for an object right away, it tries to
detect whether the value is a string, and if it is, sets that as the
object's ID. If it isn't, it then performs unmarshaling on the full
object. If that errors, we return the error.

This also has the benefit of making unmarshaling on values that are
strings faster -- we return almost immediately instead of allocating an
object, trying to decode it, failing, and falling back.

I was hoping to be able to cut down the length of the implementation by
quite a bit, and although it is shortened some, it's not too different.
I'll also try to use this opportunity to standardize on conventions
(name values `v`) and add some tests.